Returned `getIsQualityLower` to the desktop functions script
